Convert the Ripple transition and break out various classes to separate files
